fix(alerts): match scheduled alerts' preview layout for anomaly alerts
The right rail's Preview card holds the DATA preview for scheduled alerts
but held the generated SQL for anomaly alerts, with the anomaly chart
buried at the bottom of the left-hand form. Swap them so both alert types
put the same thing in the same place: the chart moves to the rail (new
AnomalyDataPreview), the SQL moves into the form as a read-only row.

Drop the Load data button — the chart now populates itself on mount and
600ms after a config edit, the way PreviewAlert already does. Firing
without a click makes half-typed configs reachable, so the gate is
stricter than the old click path needed: a detection function without its
field would have run avg(*), and a cleared number input emits "" which ??
does not catch, producing histogram(_timestamp, 'm') and a zero-width
time range. Identical rebuilds are also deduped, since an empty filter row
changes config.filters without changing the query.

The SQL row is hidden in custom_sql mode, where the user's own editor is
already on the form. Five locale keys orphaned by the button's removal are
deleted across all 15 locales and recorded in the parity spec.
