+++ web/src/components/alerts/AddAlert.vue
@@ -407,7 +407,11 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
               </div>
 
               <div v-show="activeTab === 'anomaly-config'" data-tab-pane="anomaly-config">
-                <AnomalyDetectionConfig ref="anomalyStep2Ref" :config="anomalyConfig" />
+                <AnomalyDetectionConfig
+                  ref="anomalyStep2Ref"
+                  :config="anomalyConfig"
+                  :preview-sql="anomalyPreviewSql"
+                />
               </div>
 
               <div v-show="activeTab === 'anomaly-alerting'" data-tab-pane="anomaly-alerting">
@@ -459,9 +463,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
             <div
               class="border-border-default flex shrink-0 items-center gap-2 border-b px-3 py-2.5 select-none"
             >
-              <span class="text-sm font-medium">{{
-                isAnomalyMode ? t("alerts.sqlPreview") : t("alerts.preview")
-              }}</span>
+              <span class="text-sm font-medium">{{ t("alerts.preview") }}</span>
               <template v-if="!isAnomalyMode && activeEvaluationStatus">
                 <div class="bg-border-default h-4 w-px" />
                 <OIcon
@@ -488,17 +490,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
             </div>
             <div class="min-h-0 flex-1 overflow-hidden">
               <template v-if="isAnomalyMode">
-                <!-- editor-height is QueryEditor's own API for this; a class cannot
-                   win against the inline height its rootStyle always sets. -->
-                <QueryEditor
-                  editor-id="anomaly-sql-preview"
-                  language="sql"
-                  :read-only="true"
-                  :show-auto-complete="false"
-                  :hide-nl-toggle="true"
-                  :query="anomalyPreviewSql"
-                  editor-height="100%"
-                />
+                <AnomalyDataPreview :config="anomalyConfig" />
               </template>
               <template v-else>
                 <div
@@ -597,9 +589,9 @@ import InlineSelectFolderDropdown from "../common/sidebar/InlineSelectFolderDrop
 import PreviewAlert from "./PreviewAlert.vue";
 import AlertSummary from "./AlertSummary.vue";
 import AnomalyDetectionConfig from "@/components/anomaly_detection/steps/AnomalyDetectionConfig.vue";
+import AnomalyDataPreview from "@/components/anomaly_detection/AnomalyDataPreview.vue";
 import AnomalyAlerting from "@/components/anomaly_detection/steps/AnomalyAlerting.vue";
 import AnomalySummary from "@/components/anomaly_detection/AnomalySummary.vue";
-import QueryEditor from "@/components/QueryEditor.vue";
 import { useAlertForm, defaultAlertValue } from "@/composables/useAlertForm";
 import ODrawer from "@/lib/overlay/Drawer/ODrawer.vue";
 import OTooltip from "@/lib/overlay/Tooltip/OTooltip.vue";
@@ -653,9 +645,9 @@ export default defineComponent({
     PreviewAlert,
     AlertSummary,
     AnomalyDetectionConfig,
+    AnomalyDataPreview,
     AnomalyAlerting,
     AnomalySummary,
-    QueryEditor,
     InlineSelectFolderDropdown,
     OButton,
     OToggleGroup,
