+++ web/src/locales/languages/de-DE.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Diagramme",
       "checkEvery": "Überprüfe alle",
       "checkEveryTooltip": "Wie oft der Erkennungsjob ausgeführt wird. Kann länger als der Probenzeitraum sein (z. B. alle 5 m messen, Erkennung alle 1 Stunde ausführen).",
-      "clickLoadDataHint": "Klicken Sie auf Daten laden, um eine Vorschau der Zeitreihe anzuzeigen",
-      "dataPreview": "Datenvorschau",
       "destinationRequired": "Mindestens ein Ziel ist erforderlich!",
       "detectionCharts": "Erkennungsergebnisse",
       "detectionResolution": "Auflösung der Erkennung",
@@ -7202,13 +7200,11 @@
       "disabled": "Deaktiviert",
       "disabledNotificationsInfo": "Die Anomalieerkennung wird weiterhin ausgeführt und die Ergebnisse werden in den _anomalies-Stream geschrieben. Aktivieren Sie Benachrichtigungen, um Benachrichtigungen zu erhalten, wenn Anomalien erkannt werden.",
       "enabled": "Aktiviert",
-      "enterSqlFirst": "geben Sie zuerst eine SQL-Abfrage ein",
       "fieldPlaceholder": "Feld",
       "fillInSetupStepToSeeSummary": "Füllen Sie den Einrichtungsschritt aus, um eine Zusammenfassung zu sehen.",
       "filters": "Filter",
       "lastError": "Letzter Fehler",
       "lastTrained": "Zuletzt trainiert",
-      "loadData": "Daten laden",
       "lookBackWindow": "Fenster „Blick zurück“",
       "lookBackWindowTooltip": "Gibt an, wie weit die einzelnen Erkennungsvorgänge zurückliegen. Verschließt das Fenster auch nach einer langen Pause. Die Standardeinstellung ist Jedes Intervall überprüfen.",
       "metricChart": "Metrik und Anomalien",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Nach unten scrollen",
       "seasonalityWeekly": "Stunde + Wochentag",
       "selectStreamFirst": "Wähle zuerst einen Stream",
-      "selectStreamFirstTooltip": "wähle zuerst einen Stream",
       "sensitivityAggressive": "Aggressiv",
       "sensitivityBalanced": "Ausgewogen",
       "sensitivityConservative": "Konservativ",
