+++ web/src/locales/languages/en-US.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Charts",
       "checkEvery": "Check Every",
       "checkEveryTooltip": "How often the detection job runs. Can be larger than the sample period (e.g. sample every 5m, run detection every 1h).",
-      "clickLoadDataHint": "Click Load Data to preview the time series",
-      "dataPreview": "Data preview",
       "destinationRequired": "At least one destination is required!",
       "detectionCharts": "Detection results",
       "detectionResolution": "Detection Resolution",
@@ -7202,13 +7200,11 @@
       "disabled": "Disabled",
       "disabledNotificationsInfo": "Anomaly detection will still run and write results to the _anomalies stream. Enable notifications to receive alerts when anomalies are detected.",
       "enabled": "Enabled",
-      "enterSqlFirst": "enter a SQL query first",
       "fieldPlaceholder": "Field",
       "fillInSetupStepToSeeSummary": "Fill in the setup step to see a summary.",
       "filters": "Filters",
       "lastError": "Last error",
       "lastTrained": "Last trained",
-      "loadData": "Load data",
       "lookBackWindow": "Look Back Window",
       "lookBackWindowTooltip": "How far back each detection run queries. Caps the window even after a long pause. Defaults to Check Every interval.",
       "metricChart": "Metric and anomalies",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Scroll to bottom",
       "seasonalityWeekly": "hour + day-of-week",
       "selectStreamFirst": "Select a stream first",
-      "selectStreamFirstTooltip": "select a stream first",
       "sensitivityAggressive": "Aggressive",
       "sensitivityBalanced": "Balanced",
       "sensitivityConservative": "Conservative",
