+++ web/src/locales/languages/es-ES.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Gráficos",
       "checkEvery": "Compruebe cada",
       "checkEveryTooltip": "La frecuencia con la que se ejecuta el trabajo de detección. Puede ser mayor que el período de muestreo (por ejemplo, tomar una muestra cada 5 minutos, realizar la detección cada 1 hora).",
-      "clickLoadDataHint": "Haga clic en Cargar datos para obtener una vista previa de la serie temporal",
-      "dataPreview": "Vista previa de los datos",
       "destinationRequired": "¡Se requiere al menos un destino!",
       "detectionCharts": "Resultados de detección",
       "detectionResolution": "Resolución de detección",
@@ -7202,13 +7200,11 @@
       "disabled": "Discapacitado",
       "disabledNotificationsInfo": "La detección de anomalías seguirá ejecutándose y escribirá los resultados en la secuencia _anomalies. Habilite las notificaciones para recibir alertas cuando se detecten anomalías.",
       "enabled": "Habilitado",
-      "enterSqlFirst": "introduzca primero una consulta SQL",
       "fieldPlaceholder": "Campo",
       "fillInSetupStepToSeeSummary": "Complete el paso de configuración para ver un resumen.",
       "filters": "Filtros",
       "lastError": "Último error",
       "lastTrained": "Último entrenamiento",
-      "loadData": "Cargar datos",
       "lookBackWindow": "Mirar hacia atrás",
       "lookBackWindowTooltip": "La antigüedad de cada detección en la que se ejecutan las consultas. Tapa la ventana incluso después de una pausa larga. El valor predeterminado es Comprobar todos los intervalos.",
       "metricChart": "Métrica y anomalías",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Ir al final",
       "seasonalityWeekly": "hora + día de la semana",
       "selectStreamFirst": "Selecciona primero una transmisión",
-      "selectStreamFirstTooltip": "selecciona primero una transmisión",
       "sensitivityAggressive": "Agresivo",
       "sensitivityBalanced": "Equilibrado",
       "sensitivityConservative": "Conservador",
