+++ web/src/locales/languages/fr-FR.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Graphiques",
       "checkEvery": "Vérifiez chaque",
       "checkEveryTooltip": "Fréquence d'exécution de la tâche de détection. Peut être plus longue que la période d'échantillonnage (par exemple, échantillonnage tous les 5 m, exécution de la détection toutes les 1 h).",
-      "clickLoadDataHint": "Cliquez sur Charger les données pour prévisualiser la série chronologique",
-      "dataPreview": "Aperçu des données",
       "destinationRequired": "Au moins une destination est requise !",
       "detectionCharts": "Résultats de détection",
       "detectionResolution": "Résolution de détection",
@@ -7202,13 +7200,11 @@
       "disabled": "Désactivé",
       "disabledNotificationsInfo": "La détection des anomalies sera toujours exécutée et les résultats seront enregistrés dans le flux _anomalies. Activez les notifications pour recevoir des alertes lorsque des anomalies sont détectées.",
       "enabled": "Activé",
-      "enterSqlFirst": "entrez d'abord une requête SQL",
       "fieldPlaceholder": "Champ",
       "fillInSetupStepToSeeSummary": "Remplissez l'étape de configuration pour voir un résumé.",
       "filters": "Filtres",
       "lastError": "Dernière erreur",
       "lastTrained": "Dernier entraînement",
-      "loadData": "Charger les données",
       "lookBackWindow": "Fenêtre Look Back",
       "lookBackWindowTooltip": "Depuis combien de temps chaque détection exécute les requêtes. Permet de fermer la fenêtre même après une longue pause. La valeur par défaut est Check Every Interval.",
       "metricChart": "Métrique et anomalies",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Faire défiler jusqu'en bas",
       "seasonalityWeekly": "heure + jour de la semaine",
       "selectStreamFirst": "Sélectionnez d'abord un flux",
-      "selectStreamFirstTooltip": "sélectionnez d'abord un flux",
       "sensitivityAggressive": "Agressif",
       "sensitivityBalanced": "Équilibré",
       "sensitivityConservative": "Conservateur",
