+++ web/src/locales/languages/it-IT.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Grafici",
       "checkEvery": "Controlla ogni",
       "checkEveryTooltip": "Con quale frequenza viene eseguito il processo di rilevamento. Può essere superiore al periodo di campionamento (ad esempio campionamento ogni 5 m, rilevamento eseguito ogni 1 ora).",
-      "clickLoadDataHint": "Fai clic su Carica dati per visualizzare l'anteprima delle serie temporali",
-      "dataPreview": "Anteprima dei dati",
       "destinationRequired": "È richiesta almeno una destinazione!",
       "detectionCharts": "Risultati del rilevamento",
       "detectionResolution": "Risoluzione di rilevamento",
@@ -7202,13 +7200,11 @@
       "disabled": "Disabili",
       "disabledNotificationsInfo": "Il rilevamento delle anomalie verrà comunque eseguito e scriverà i risultati nel flusso _anomalies. Abilita le notifiche per ricevere avvisi quando vengono rilevate anomalie.",
       "enabled": "Abilitato",
-      "enterSqlFirst": "inserisci prima una query SQL",
       "fieldPlaceholder": "Campo",
       "fillInSetupStepToSeeSummary": "Compila il passaggio di configurazione per vedere un riepilogo.",
       "filters": "Filtri",
       "lastError": "Ultimo errore",
       "lastTrained": "Ultimo addestramento",
-      "loadData": "Carica dati",
       "lookBackWindow": "Finestra di lookback",
       "lookBackWindowTooltip": "Quanto tempo fa ogni rilevamento ha eseguito le interrogazioni. Chiude la finestra anche dopo una lunga pausa. L'impostazione predefinita è Check Every interval.",
       "metricChart": "Metrica e anomalie",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Scorri fino in fondo",
       "seasonalityWeekly": "ora + giorno della settimana",
       "selectStreamFirst": "Seleziona prima uno stream",
-      "selectStreamFirstTooltip": "seleziona prima uno stream",
       "sensitivityAggressive": "Aggressivo",
       "sensitivityBalanced": "Bilanciato",
       "sensitivityConservative": "Conservativo",
