+++ web/src/locales/languages/ja-JP.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "グラフ",
       "checkEvery": "すべてチェック",
       "checkEveryTooltip": "検出ジョブの実行頻度。サンプル期間より長くてもかまいません (例:5 m ごとにサンプリングし、1 時間ごとに検出を実行する)。",
-      "clickLoadDataHint": "[データをロード] をクリックして時系列をプレビューします",
-      "dataPreview": "データプレビュー",
       "destinationRequired": "宛先が少なくとも1つ必要です！",
       "detectionCharts": "検出結果",
       "detectionResolution": "検出解像度",
@@ -7202,13 +7200,11 @@
       "disabled": "無効",
       "disabledNotificationsInfo": "異常検出は引き続き実行され、結果が _anomalies ストリームに書き込まれます。異常が検出されたときにアラートを受け取るように通知を有効にします。",
       "enabled": "有効",
-      "enterSqlFirst": "最初に SQL クエリを入力してください",
       "fieldPlaceholder": "フィールド",
       "fillInSetupStepToSeeSummary": "セットアップ手順を入力するとサマリーが表示されます。",
       "filters": "フィルター",
       "lastError": "最終エラー",
       "lastTrained": "最終トレーニング",
-      "loadData": "データを読み込む",
       "lookBackWindow": "ルックバックウィンドウ",
       "lookBackWindowTooltip": "各検出でクエリが実行されるまでの時間。しばらく待ってもウィンドウをキャップします。デフォルトは [間隔ごとにチェック] です。",
       "metricChart": "メトリクスと異常",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "一番下までスクロール",
       "seasonalityWeekly": "時間+曜日",
       "selectStreamFirst": "最初にストリームを選択してください",
-      "selectStreamFirstTooltip": "最初にストリームを選択してください",
       "sensitivityAggressive": "積極的",
       "sensitivityBalanced": "バランス",
       "sensitivityConservative": "保守的",
