+++ web/src/locales/languages/ko-KR.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "차트",
       "checkEvery": "모두 확인",
       "checkEveryTooltip": "탐지 작업의 실행 빈도샘플 기간보다 길 수 있습니다 (예: 5m마다 샘플링, 1시간마다 탐지 실행).",
-      "clickLoadDataHint": "데이터 로드를 클릭하여 시계열을 미리 봅니다.",
-      "dataPreview": "데이터 미리보기",
       "destinationRequired": "대상이 하나 이상 필요합니다!",
       "detectionCharts": "탐지 결과",
       "detectionResolution": "감지 해상도",
@@ -7202,13 +7200,11 @@
       "disabled": "장애인",
       "disabledNotificationsInfo": "예외 항목 탐지는 계속 실행되고 결과를 _anomalies 스트림에 기록합니다.이상 징후가 감지되면 알림을 받을 수 있도록 하세요.",
       "enabled": "활성화됨",
-      "enterSqlFirst": "먼저 SQL 쿼리를 입력합니다.",
       "fieldPlaceholder": "필드",
       "fillInSetupStepToSeeSummary": "요약을 보려면 설정 단계를 입력하세요.",
       "filters": "필터",
       "lastError": "마지막 오류",
       "lastTrained": "마지막 학습",
-      "loadData": "데이터 로드",
       "lookBackWindow": "룩 백 윈도우",
       "lookBackWindowTooltip": "각 탐지에서 쿼리를 실행한 시점잠시 멈춘 후에도 창에 캡이 표시됩니다.기본값은 [모든 간격 확인] 입니다.",
       "metricChart": "메트릭 및 이상",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "맨 아래로 스크롤",
       "seasonalityWeekly": "시간+요일",
       "selectStreamFirst": "먼저 스트림을 선택하세요",
-      "selectStreamFirstTooltip": "먼저 스트림을 선택하세요",
       "sensitivityAggressive": "공격적",
       "sensitivityBalanced": "균형",
       "sensitivityConservative": "보수적",
