+++ web/src/locales/languages/nl-NL.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Grafieken",
       "checkEvery": "Controleer elke",
       "checkEveryTooltip": "Hoe vaak de detectieopdracht wordt uitgevoerd. Kan groter zijn dan de steekproefperiode (bijv. monster elke 5 meter, detectie elke 1 uur uitvoeren).",
-      "clickLoadDataHint": "Klik op Gegevens laden om een voorbeeld van de tijdreeks te bekijken",
-      "dataPreview": "Voorbeeld van de gegevens",
       "destinationRequired": "Minstens één bestemming is vereist!",
       "detectionCharts": "Detectieresultaten",
       "detectionResolution": "Resolutie van de detectie",
@@ -7202,13 +7200,11 @@
       "disabled": "Uitgeschakeld",
       "disabledNotificationsInfo": "De detectie van afwijkingen wordt nog steeds uitgevoerd en de resultaten worden naar de _anomalieënstream geschreven. Schakel meldingen in om waarschuwingen te ontvangen wanneer er afwijkingen worden gedetecteerd.",
       "enabled": "Ingeschakeld",
-      "enterSqlFirst": "voer eerst een SQL-query in",
       "fieldPlaceholder": "Veld",
       "fillInSetupStepToSeeSummary": "Vul de setupstap in om een samenvatting te zien.",
       "filters": "Filters",
       "lastError": "Laatste fout",
       "lastTrained": "Laatst getraind",
-      "loadData": "Gegevens laden",
       "lookBackWindow": "Kijk terug naar het venster",
       "lookBackWindowTooltip": "Hoe ver terug elke detectie voert query's uit. Sluit het venster af, zelfs na een lange pauze. De standaardinstelling is om elk interval te controleren.",
       "metricChart": "Metriek en anomalieën",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Scroll naar beneden",
       "seasonalityWeekly": "uur + dag van de week",
       "selectStreamFirst": "Selecteer eerst een stream",
-      "selectStreamFirstTooltip": "selecteer eerst een stream",
       "sensitivityAggressive": "Agressief",
       "sensitivityBalanced": "Gebalanceerd",
       "sensitivityConservative": "Conservatief",
