+++ web/src/locales/languages/pl-PL.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Wykresy",
       "checkEvery": "Sprawdzaj co",
       "checkEveryTooltip": "Jak często uruchamiane jest zadanie wykrywania. Może być większe niż okres próbkowania (np. próbkowanie co 5 min, wykrywanie co 1 godz.).",
-      "clickLoadDataHint": "Kliknij Wczytaj dane, aby wyświetlić podgląd szeregu czasowego",
-      "dataPreview": "Podgląd danych",
       "destinationRequired": "Co najmniej jeden cel jest wymagany!",
       "detectionCharts": "Wyniki wykrywania",
       "detectionResolution": "Rozdzielczość wykrywania",
@@ -7202,13 +7200,11 @@
       "disabled": "Wyłączone",
       "disabledNotificationsInfo": "Wykrywanie anomalii będzie nadal działać i zapisywać wyniki do strumienia _anomalies. Włącz powiadomienia, aby otrzymywać alerty po wykryciu anomalii.",
       "enabled": "Włączone",
-      "enterSqlFirst": "najpierw wprowadź zapytanie SQL",
       "fieldPlaceholder": "Pole",
       "fillInSetupStepToSeeSummary": "Wypełnij krok konfiguracji, aby zobaczyć podsumowanie.",
       "filters": "Filtry",
       "lastError": "Ostatni błąd",
       "lastTrained": "Ostatnio trenowany",
-      "loadData": "Wczytaj dane",
       "lookBackWindow": "Okno wsteczne",
       "lookBackWindowTooltip": "Jak daleko wstecz sięga każde zapytanie wykrywania. Ogranicza okno nawet po długiej przerwie. Domyślnie równa się interwałowi sprawdzania.",
       "metricChart": "Metryka i anomalie",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Przewiń do dołu",
       "seasonalityWeekly": "godzina + dzień tygodnia",
       "selectStreamFirst": "Najpierw wybierz strumień",
-      "selectStreamFirstTooltip": "najpierw wybierz strumień",
       "sensitivityAggressive": "Agresywna",
       "sensitivityBalanced": "Zrównoważona",
       "sensitivityConservative": "Konserwatywna",
