+++ web/src/locales/languages/pt-PT.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Gráficos",
       "checkEvery": "Verifique cada",
       "checkEveryTooltip": "Com que frequência o trabalho de detecção é executado. Pode ser maior do que o período de amostragem (por exemplo, amostra a cada 5m, detecção de execução a cada 1h).",
-      "clickLoadDataHint": "Clique em Carregar dados para visualizar a série temporal",
-      "dataPreview": "Pré-visualização dos dados",
       "destinationRequired": "Pelo menos um destino é obrigatório!",
       "detectionCharts": "Resultados de deteção",
       "detectionResolution": "Resolução de detecção",
@@ -7202,13 +7200,11 @@
       "disabled": "Desativado",
       "disabledNotificationsInfo": "A detecção de anomalias ainda será executada e gravará os resultados no fluxo _anomalies. Ative as notificações para receber alertas quando forem detectadas anomalias.",
       "enabled": "Ativado",
-      "enterSqlFirst": "insira uma consulta SQL primeiro",
       "fieldPlaceholder": "Campo",
       "fillInSetupStepToSeeSummary": "Preencha a etapa de configuração para ver um resumo.",
       "filters": "Filtros",
       "lastError": "Último erro",
       "lastTrained": "Último treino",
-      "loadData": "Carregar dados",
       "lookBackWindow": "Janela Olhe para trás",
       "lookBackWindowTooltip": "Há quanto tempo cada detecção executa consultas. Fecha a janela mesmo depois de uma longa pausa. O padrão é Verificar cada intervalo.",
       "metricChart": "Métrica e anomalias",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Rolar até o final",
       "seasonalityWeekly": "hora+ dia da semana",
       "selectStreamFirst": "Selecione um stream primeiro",
-      "selectStreamFirstTooltip": "selecione um stream primeiro",
       "sensitivityAggressive": "Agressivo",
       "sensitivityBalanced": "Equilibrado",
       "sensitivityConservative": "Conservador",
