+++ web/src/locales/languages/ru-RU.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Графики",
       "checkEvery": "Проверять каждые",
       "checkEveryTooltip": "Как часто выполняется задание обнаружения. Может быть больше периода выборки (например, выборка каждые 5 мин, обнаружение каждый час).",
-      "clickLoadDataHint": "Нажмите «Загрузить данные» для предварительного просмотра временного ряда",
-      "dataPreview": "Предпросмотр данных",
       "destinationRequired": "Требуется хотя бы один получатель!",
       "detectionCharts": "Результаты обнаружения",
       "detectionResolution": "Разрешение обнаружения",
@@ -7202,13 +7200,11 @@
       "disabled": "Отключено",
       "disabledNotificationsInfo": "Обнаружение аномалий будет по-прежнему выполняться и записывать результаты в поток _anomalies. Включите уведомления для получения оповещений при обнаружении аномалий.",
       "enabled": "Включено",
-      "enterSqlFirst": "сначала введите SQL-запрос",
       "fieldPlaceholder": "Поле",
       "fillInSetupStepToSeeSummary": "Заполните шаг настройки, чтобы увидеть сводку.",
       "filters": "Фильтры",
       "lastError": "Последняя ошибка",
       "lastTrained": "Последнее обучение",
-      "loadData": "Загрузить данные",
       "lookBackWindow": "Окно просмотра назад",
       "lookBackWindowTooltip": "Как далеко назад запрашивает каждое выполнение обнаружения. Ограничивает окно даже после длительной паузы. По умолчанию равно интервалу проверки.",
       "metricChart": "Метрика и аномалии",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Прокрутить вниз",
       "seasonalityWeekly": "час + день недели",
       "selectStreamFirst": "Сначала выберите поток",
-      "selectStreamFirstTooltip": "сначала выберите поток",
       "sensitivityAggressive": "Агрессивная",
       "sensitivityBalanced": "Сбалансированная",
       "sensitivityConservative": "Консервативная",
