+++ web/src/locales/languages/tr-TR.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Grafikler",
       "checkEvery": "Her birini kontrol et",
       "checkEveryTooltip": "Algılama işinin ne sıklıkta çalıştığı. Numune periyodundan daha büyük olabilir (örn. her 5m'de bir numune, her 1 saatte bir algılama çalıştırın).",
-      "clickLoadDataHint": "Zaman serilerini önizlemek için Veri Yükle'yi tıklatın",
-      "dataPreview": "Veri önizlemesi",
       "destinationRequired": "En az bir hedef gereklidir!",
       "detectionCharts": "Algılama sonuçları",
       "detectionResolution": "Algılama Çözünürlüğü",
@@ -7202,13 +7200,11 @@
       "disabled": "Engelli",
       "disabledNotificationsInfo": "Anomali algılama çalışmaya devam edecek ve sonuçları _anomaliler akışına yazacaktır. Anormallikler tespit edildiğinde uyarı almak için bildirimleri etkinleştirin.",
       "enabled": "Etkin",
-      "enterSqlFirst": "önce bir SQL sorgusu girin",
       "fieldPlaceholder": "Tarla",
       "fillInSetupStepToSeeSummary": "Özeti görmek için kurulum adımını doldurun.",
       "filters": "Filtreler",
       "lastError": "Son hata",
       "lastTrained": "Son eğitim",
-      "loadData": "Veri yükle",
       "lookBackWindow": "Pencereye Bakın",
       "lookBackWindowTooltip": "Her algılamanın ne kadar geriye dönük çalıştırma sorguları. Uzun bir duraklamadan sonra bile pencereyi kapatır. Varsayılan olarak Her Aralığı Kontrol Et.",
       "metricChart": "Metrik ve anomaliler",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "En alta kaydır",
       "seasonalityWeekly": "saat+haftanın günü",
       "selectStreamFirst": "Önce bir yayın seçin",
-      "selectStreamFirstTooltip": "önce bir akış seçin",
       "sensitivityAggressive": "Agresif",
       "sensitivityBalanced": "Dengeli",
       "sensitivityConservative": "Muhafazakâr",
