+++ web/src/locales/languages/vi-VN.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "Biểu đồ",
       "checkEvery": "Kiểm tra mỗi",
       "checkEveryTooltip": "Tần suất chạy tác vụ phát hiện. Có thể lớn hơn chu kỳ lấy mẫu (ví dụ: lấy mẫu mỗi 5 phút, chạy phát hiện mỗi 1 giờ).",
-      "clickLoadDataHint": "Nhấp vào Tải dữ liệu để xem trước chuỗi thời gian",
-      "dataPreview": "Xem trước dữ liệu",
       "destinationRequired": "Cần ít nhất một đích!",
       "detectionCharts": "Kết quả phát hiện",
       "detectionResolution": "Độ phân giải phát hiện",
@@ -7202,13 +7200,11 @@
       "disabled": "Đã tắt",
       "disabledNotificationsInfo": "Phát hiện bất thường vẫn sẽ chạy và ghi kết quả vào luồng _anomalies. Bật thông báo để nhận cảnh báo khi phát hiện bất thường.",
       "enabled": "Đã bật",
-      "enterSqlFirst": "nhập truy vấn SQL trước",
       "fieldPlaceholder": "Trường",
       "fillInSetupStepToSeeSummary": "Điền vào bước thiết lập để xem tóm tắt.",
       "filters": "Bộ lọc",
       "lastError": "Lỗi cuối cùng",
       "lastTrained": "Lần huấn luyện cuối",
-      "loadData": "Tải dữ liệu",
       "lookBackWindow": "Cửa sổ nhìn lại",
       "lookBackWindowTooltip": "Khoảng thời gian tối đa mà mỗi lần chạy phát hiện truy vấn. Giới hạn cửa sổ ngay cả sau khi tạm dừng lâu. Mặc định là khoảng thời gian Kiểm tra mỗi lần.",
       "metricChart": "Chỉ số và bất thường",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "Cuộn xuống cuối",
       "seasonalityWeekly": "giờ + ngày trong tuần",
       "selectStreamFirst": "Chọn một luồng trước",
-      "selectStreamFirstTooltip": "chọn một luồng trước",
       "sensitivityAggressive": "Quyết liệt",
       "sensitivityBalanced": "Cân bằng",
       "sensitivityConservative": "Thận trọng",
