+++ web/src/locales/languages/zh-CN.json
@@ -7191,8 +7191,6 @@
       "chartsTab": "图表",
       "checkEvery": "检查每个",
       "checkEveryTooltip": "检测任务的运行频率。可以大于采样周期（例如，每 5 分钟采样一次，每 1 小时运行一次检测）。",
-      "clickLoadDataHint": "单击 “加载数据” 预览时间序列",
-      "dataPreview": "数据预览",
       "destinationRequired": "至少需要一个目标！",
       "detectionCharts": "检测结果",
       "detectionResolution": "探测分辨率",
@@ -7202,13 +7200,11 @@
       "disabled": "已禁用",
       "disabledNotificationsInfo": "异常检测仍会运行并将结果写入 _anomalies 数据流。启用通知以便在检测到异常时收到告警。",
       "enabled": "已启用",
-      "enterSqlFirst": "先输入 SQL 查询",
       "fieldPlaceholder": "字段",
       "fillInSetupStepToSeeSummary": "填写设置步骤以查看摘要。",
       "filters": "过滤器",
       "lastError": "最近错误",
       "lastTrained": "上次训练",
-      "loadData": "加载数据",
       "lookBackWindow": "向后看窗口",
       "lookBackWindowTooltip": "每次检测运行查询的时间可以追溯到多久以前。即使停顿了很长时间，也要关闭窗口。默认为 “每隔一段时间都检查”。",
       "metricChart": "指标与异常",
@@ -7233,7 +7229,6 @@
       "scrollToBottom": "滚动到底部",
       "seasonalityWeekly": "一小时 + 一周中的某一天",
       "selectStreamFirst": "先选择一个数据流",
-      "selectStreamFirstTooltip": "先选择一个数据流",
       "sensitivityAggressive": "激进",
       "sensitivityBalanced": "平衡",
       "sensitivityConservative": "保守",
