+++ web/src/locales/languages/zh-TW.json
@@ -7186,8 +7186,6 @@
       "chartsTab": "圖表",
       "checkEvery": "檢查頻率",
       "checkEveryTooltip": "偵測工作的執行頻率。可以大於取樣週期 (例如每 5 分鐘取樣，每 1 小時執行偵測)。",
-      "clickLoadDataHint": "按一下 [載入資料] 以預覽時間序列",
-      "dataPreview": "資料預覽",
       "destinationRequired": "至少需要一個目標！",
       "detectionCharts": "偵測結果",
       "detectionResolution": "偵測解析度",
@@ -7197,13 +7195,11 @@
       "disabled": "已停用",
       "disabledNotificationsInfo": "異常偵測仍會執行並將結果寫入 _anomalies 串流。啟用通知以在偵測到異常時收到警示。",
       "enabled": "已啟用",
-      "enterSqlFirst": "請先輸入 SQL 查詢",
       "fieldPlaceholder": "欄位",
       "fillInSetupStepToSeeSummary": "填寫設定步驟以查看摘要。",
       "filters": "篩選條件",
       "lastError": "最近錯誤",
       "lastTrained": "上次訓練",
-      "loadData": "載入資料",
       "lookBackWindow": "回溯視窗",
       "lookBackWindowTooltip": "每次偵測執行查詢的回溯範圍。即使暫停很長時間也會限制視窗。預設為檢查頻率間隔。",
       "metricChart": "指標與異常",
@@ -7228,7 +7224,6 @@
       "scrollToBottom": "捲動至底部",
       "seasonalityWeekly": "小時 + 星期幾",
       "selectStreamFirst": "請先選擇串流",
-      "selectStreamFirstTooltip": "請先選擇串流",
       "sensitivityAggressive": "積極",
       "sensitivityBalanced": "平衡",
       "sensitivityConservative": "保守",
